requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
tensorflow==2.15.0
keras==2.15.0
nltk==3.8.1
scikit-learn==1.6.1
numpy==1.24.3
huggingface-hub==0.20.3
python-dotenv==1.0.0
ml-dtypes==0.2.0
joblib==1.5.1
transformers==4.35.2